Basic skeleton.
The antlr runtime will be fetched and unpacked by the Makefile as
necessary to avoid installation.

The first frontend grammar is an outline of C - it compiles but it
is currently just a vague sketch of the top-level language parts.

The antlr grammar is rewritten to insert names and targets as there
are no command-line switches for this. Currently C and Python target
parsers are generated.