Fixed legion errors to get the C target working on OSX.
Needed to hack around the missing configure script to get the runtime
library to build, sketch out some code that uses the C API and hide
the hidden channel in the grammar behind a sed replacement as the
syntax differs between C and Python.

This version runs the trivial demo in both languages on OSX. The steps
to get the runtime working are listed in the README but not automated
in the Makefile yet.