Started building the model in C.
I think this is going to be really kludgey. The psuedo-class interface
in the C target is very awkward and verbose to use. I've already tried
compiling the parser under g++ but there is no joy.

The analysers that we want to write will almost certainly be easier in
C++ as the STL is going to be very handy, and for manipulating tree
structures having real constructors and destructors will make life a lot
more simple.

The problem is giong to be the interface between some kind of shim around
the parser written in C and using malloc, and calling code written in C++
and using new/delete.