More rearranging of parsing interface.
The stream of tokens on the end of the function definition is now parsed
as a partial Decl and used to specify the return type of the function.

Next up is expanding the test case to a more reasonable set (ie the list
scraped into the python code), and handling pointers in type declarations.