Splitting Decl.
The original Decl class is now split into a Type and a Decl to smooth
over the tricky corner cases for things like parameters and prototypes.
The new models directory is going to be the IR for the project.

This version compiles and runs, the output is lacking identifiers and
array bounds.