Trying to add function pointers to the parser.
Folded up the declarator token stream without any problems, added a
special case to the parenthesised tail for the declarator.

Tried to add a new token to use as a tree node for the case... and
all hell breaks loose. Really can't see what has gone wrong, adding
any additional token to the block at the start of the parser causes
it to break quite badly, chew up the input stream and then abort
on some dodgy string processing. Need to try and diagnose what is
going on another day.