Using the source from gawk as a testcase.
The tarball gets pulled in and expanded by the Makefile.

The array.c file is the first testcase. Currently the lexer can process
the file (there are some broken parts in operator precedence to fix
later). The grammar processes the input with some errors and builds
a somewhat broken tree.

There are lots of nil nodes inserted everywhere to look into first.