Model of array.c successfully constructed.
The parameter parsing had changed a long time ago. We can no longer
use Decl::parse on function definition arguments as it now relies
on explicit dtor-structure in the form of DECL sub-nodes. This is
to separate type/symbol names clearly in declarations.

Instead there is a new function is a specialisation of partition
for the stateful problem that is splitting the token stream into
type parts and names.

Everything looks good on array, now going to try the rest of the
gawk source code to look for broken cases.