Fixed dependency bug.
There have been several issues that traced back to stale objects being linked
into the trivial demo. The cause was that the dependency for the cInCLexer
module pointed to a generated C file - but it actually gets rebuilt as part
of generating cInCParser. Added an extra build rule and most of the DECL
parsing errors have vanished.