Ambiguity resolved.
The second-pass parsing now works and correctly classifies statatements
as either actual statements, or declarations based on the results of
reparsing the stream slice and then validating the Decl object against
the current contents of the SymbolTable.

In the array testcase the two failures are NODE and size_t types that
are not in the top-level symbol-table as they should be in the headers
SymbolTable.