Massive Rewrite Part I
Started the day adding structs and unions to the data-model. I did not
want to write a bunch of useless code that parsed these into Types, just
so that they could be converted to DataTypes.

Instead I've started to rip out the old parser representation completely
as it is no longer needed. The ANTLR -> model steps are really all code
with some temporary storage - there is no intermediate representation.

This will mean that c-init is removed completely and the two remaining
modules will be c-repr and c-build. The building code (with ANTLR
dependencies) will still be separate from the final representation of
the C data-model.

Current status is that the function handling has been disabled for now,
and most of the data handling has been ported across and rewritten to
generate DataType and FuncType objects directly. The SymbolTables have
been threaded through so that this can occur in a single pass. This
leads me to believe that the lexer-hack-less parser is not really any
simpler as there is now a substantial amount of parsing going on in the
C code with access to the type-table.

This code is pure C though and operates as a simple recursive descent
tree parser on the AST, so maybe it does work out to be less complex
in the end.

This version does compile and run against the testdecls file, but only
the simple cases are handled correctly and there is a lot of bugs to
be worked out.