Massive Rewrite Finished.
Function definitions are now processed again. The type of the function
is built and registered in the symbol table. There is a bug that pointer
types in the parameters and return are not displayed as pointers when
the SymbolTable is dumped.

The statement reparsing inside function definitions is now implemented
again and a lot cleaner than before. The declarations inside the function
are already in the SymbolTable as a result of using convertDECL. There is
a stream of statements ready for further parsing.