Quick look at forward-references.
Tried allocating the record on a forward-reference and then inserting
the fields on the definition. This approach won't work as updating
the value changes the ordering, this can break uniqueness and equality
guarantees in the SymbolTable.

The other obvious approach is to rewrite DataType* when merging two
types: this is not feasible as we don't know where they may be stored ;
they are used as external identifiers in functions and expressions.

The only remaining option is a form of late-binding: remove the fields
from the datatype and introduce a separate Record class that represents
unions and structs. In the DataType class just store the tag-name, and
use the tags lookup in the SymbolTable to do late resolution.