Changes to split DataType into RecType.
This is not so much a branch as a stump - decided that this was a really
bad idea as the equality and ordering constraints that we need on a
DataType pollute the RecType. Even with late binding in the SymbolTable
there would be collision problems. Left these changes as it does simplify
the DataType class and make the model more modular - but it is not worth
the effort currently to propagate these changes through the affected parts
of c-build.