Abandoned canonical types.
Sometimes things are just not worth the hassle. After three weeks of
trying to solve this it is just simpler to accept that we will have
duplicates in the symbol table.

The kind of aliasing that needed to be avoided is when multiple
uses of records must point to the same list of declarations. This has
been solved in a very simple way by leaving the tag name in the atomic
part and using late-binding to the tags lookup in the SymbolTable.

Simple test-files now run through, some of the forward-references
are not overwriting correctly, and there is a ton of junk code to
eliminate.